# apilayer/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    MessageResponse:
        properties:
            Message:
                type: string
        title: MessageResponse ...
        type: object
        x-go-package: github.com/mohit2530/communityCare
    Notes:
        properties:
            color:
                type: string
                x-go-name: Color
            completionDate:
                format: date-time
                type: string
                x-go-name: CompletionDate
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            created_by:
                type: string
                x-go-name: CreatedBy
            creator:
                type: string
                x-go-name: Creator
            description:
                type: string
                x-go-name: Description
            location:
                $ref: '#/definitions/Location'
            noteID:
                type: string
                x-go-name: ID
            sharable_groups:
                items:
                    format: uuid
                    type: string
                type: array
                x-go-name: SharableGroups
            status:
                type: string
                x-go-name: Status
            status_description:
                type: string
                x-go-name: StatusDescription
            status_name:
                type: string
                x-go-name: StatusName
            title:
                type: string
                x-go-name: Title
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            updated_by:
                type: string
                x-go-name: UpdatedBy
            updator:
                type: string
                x-go-name: Updator
        title: Notes ...
        type: object
        x-go-name: Note
        x-go-package: github.com/mohit2530/communityCare/model
...
info:
    description: |-
        # Documentation for the Asset alert api layer.

        Scehmes: http
    title: alert Api Layer
    version: 1.0.0
paths:
    /api/profile/{id}/inventories:
        post:
            operationId: removeSelectedInventory
            parameters:
                - description: The id of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The inventory object to remove from the db
                  in: body
                  name: Inventory
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The inventory object to remove from the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Update selected inventory with details.'
            tags:
                - RemoveSelectedInventory
        put:
            operationId: updateSelectedInventory
            parameters:
                - description: The id of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The inventory object to add into the db
                  in: body
                  name: Inventory
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The inventory object to add into the db
                    type: object
            responses:
                "200":
                    description: Inventory
                    schema:
                        $ref: '#/definitions/Inventory'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Update selected inventory with details.'
            tags:
                - UpdateSelectedInventory
    /api/profile/{id}/inventories/bulk:
        post:
            operationId: addInventoryInBulk
            parameters:
                - description: The id of the selected inventory
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The list of inventories to add into the db to support bulk upload
                  in: body
                  name: InventoryListRequest
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The list of inventories to add into the db to support bulk upload
                    type: object
            responses:
                "200":
                    $ref: '#/responses/model.Inventory'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Bulk upload inventory list. returns the list of uploaded items.'
            tags:
                - AddInventoryInBulk
    /api/profile/{id}/notes:
        delete:
            operationId: removeNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Removes the note from the db
            tags:
                - RemoveNote
        get:
            operationId: getNotes
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Notes
                    schema:
                        items:
                            $ref: '#/definitions/Notes'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieves the list of notes for the selected user
            tags:
                - GetNotes
        post:
            operationId: addNewNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The note object to add into the db
                  in: body
                  name: Note
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The note object to add into the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Add a new note to the database
            tags:
                - AddNewNote
        put:
            operationId: updateNote
            parameters:
                - description: The id of the selected note
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The note object to update into the db
                  in: body
                  name: Note
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The note object to update into the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Updates an existing note in the database
            tags:
                - UpdateNote
    /api/v1/categories:
        get:
            operationId: getAllCategories
            parameters:
                - description: The userID of the selected user
                  in: query
                  name: id
                  required: true
                  type: string
                - description: The limit of categories
                  format: int32
                  in: query
                  name: limit
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/CategoryList'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieves the list of categories that each asset can be associated with. Each user can have their own set of categories. All categories are specific to the selected user.
            tags:
                - GetAllCategories
...
    /api/v1/category/{id}:
        delete:
            operationId: removeCategory
            parameters:
                - description: The id of the category to delete
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Removes a selected category based on the id
            tags:
                - RemoveCategory
        post:
            operationId: createCategory
            parameters:
                - description: The category object to save
                  in: body
                  name: Category
                  required: true
                  schema:
                    $ref: '#/definitions/Category'
                    description: The category object to save
                    type: object
            responses:
                "200":
                    description: Category
                    schema:
                        $ref: '#/definitions/Category'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Create category
            tags:
                - CreateCategory
        put:
            operationId: updateCategory
            parameters:
                - description: The category object to update details for
                  in: query
                  name: Category
                  required: true
                  type: object
            responses:
                "200":
                    description: Category
                    schema:
                        $ref: '#/definitions/Category'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Update category function updates the selected category with new values
            tags:
                - UpdateCategory
    /api/v1/category/items:
        get:
            operationId: getAllCategoryItems
            parameters:
                - description: The userID of the selected user
                  in: query
                  name: id
                  required: true
                  type: string
                - description: The limit of categories
                  format: int32
                  in: query
                  name: limit
                  required: true
                  type: integer
                - description: The category id of the selected plan
                  in: path
                  name: catID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/CategoryItems'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Retrieves the list of assets for a specific category.'
            tags:
                - GetAllCategoryItems
        post:
            operationId: addItemsInMaintenancePlan
            parameters:
                - description: The object containing the array of assets to update with the userID
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/MaintenanceItemRequest'
                    description: The object containing the array of assets to update with the userID
                    type: object
            responses:
                "200":
                    description: MaintenanceItemResponse
                    schema:
                        items:
                            $ref: '#/definitions/MaintenanceItemResponse'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Add selected items in a specific maintenance plan
            tags:
                - AddItemsInMaintenancePlan
    /api/v1/locations:
        get:
            operationId: getAllStorageLocations
            responses:
                "200":
                    description: StorageLocation
                    schema:
                        items:
                            $ref: '#/definitions/StorageLocation'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Retrieves the list of locations that users can use to store an asset.'
            tags:
                - GetAllStorageLocations
    /api/v1/logout:
        post:
            operationId: logout
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Logs users out of the database system.'
            tags:
                - Logout
    /api/v1/maintenance-plans:
        get:
            description: Users can assign asset to multiple plans.
            operationId: getAllMaintenancePlans
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The limit of maintenance plans
                  format: int32
                  in: path
                  name: limit
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/MaintenancePlan'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: |-
                Retrieves the list of maintenance plans that each asset can be associated with.
                Each user can have thier own set of maintenance plans. All plans are specific to the selected user
            tags:
                - GetAllMaintenancePlans
    /api/v1/plan/:
        get:
            operationId: GetMaintenancePlan
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The maintenance id of the selected plan
                  in: path
                  name: mID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/MaintenancePlan'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Retrieve a selected maintenance plan
            tags:
                - GetMaintenancePlan
    /api/v1/plan/{id}:
        delete:
            operationId: removeMaintenancePlan
            parameters:
                - description: The id of the maintenance plan to delete
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Removes a selected maintenance plan based on the id
            tags:
                - RemoveMaintenancePlan
        post:
            operationId: createMaintenancePlan
            parameters:
                - description: The object containing details of the maintenance plan
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/MaintenancePlan'
                    description: The object containing details of the maintenance plan
                    type: object
            responses:
                "200":
                    $ref: '#/responses/MaintenancePlan'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Create maintenance plan
            tags:
                - CreateMaintenancePlan
        put:
            operationId: UpdateMaintenancePlan
            parameters:
                - description: The object containing details of the maintenance plan
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/MaintenancePlan'
                    description: The object containing details of the maintenance plan
                    type: object
            responses:
                "200":
                    $ref: '#/responses/MaintenancePlan'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: Update maintenance plan function updates the selected maintenance plan with new values
            tags:
                - UpdateMaintenancePlan
    /api/v1/plans/items:
        get:
            operationId: getAllMaintenancePlanItems
            parameters:
                - description: The userID of the selected user
                  in: query
                  name: id
                  required: true
                  type: string
                - description: The limit of maintenance plans
                  format: int32
                  in: query
                  name: limit
                  required: true
                  type: integer
                - description: The maintenance id of the selected plan
                  in: path
                  name: mID
                  required: true
                  type: string
            responses:
                "200":
                    description: MaintenanceItemResponse
                    schema:
                        items:
                            $ref: '#/definitions/MaintenanceItemResponse'
                        type: array
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Retrieves the list of assets for a specific maintenance plans.'
            tags:
                - GetAllMaintenancePlanItems
...
produces:
    - application/json
swagger: "2.0"